Get basic read behavior working
This can now take an arbitrary JSONAPI success response and turn it into
a graph of objects. Currently we are being fairly restrictive in what we
accept. Will have to add more robust handling down the line, but this
should work with any successful Graphiti response at least, if not all
jsonapi documents.